+++ NEWS.md
@@ -15,6 +15,7 @@
 - Added `distr_categorical()` (#576)
 - Added `distr_mixture_same_family()` (#576)
 - Tensors in the R side that point to the same C++ Tensor are now guaranteed to be the same object. This allows to easily determine unique model parameters. (#582)
+- Fixed bug in the `log_prob` of `distr_bernoulli()` (#581)
 
 # torch 0.3.0